+++ src/main/java/domain/player/participant/betresult/resultstate/WinState.java
@@ -1,4 +1,4 @@
-package domain.player.participant.betresult;
+package domain.player.participant.betresult.resultstate;
 
 import domain.player.participant.Money;
